requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
requests==2.20.0
six==1.11.0
urllib3==1.24
websocket-client==0.53.0
